+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 micro-base is in a pre-1.0 state. This means that its APIs and behavior are subject to breaking changes without deprecation notices. Until 1.0, version numbers will follow a [Semver][]-ish `0.y.z` format, where `y` is incremented when new features or breaking changes are introduced, and `z` is incremented for lesser changes or bug fixes.
 
+## [0.3.1][] (2016-05-30)
+
+* JWT token verification. 
+* Token pass to downstream services. 
+* CID creation.
+
 ## [0.3.0][] (2016-05-26)
 
 * Added docker examples